pet/requirements.txt
alembic==1.7.4
amqp==5.0.6
autopep8==1.6.0
billiard==3.6.4.0
blinker==1.4
celery==5.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
decorator==5.1.0
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Dance==5.1.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
GitHub-Flask==3.2.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
infinity==1.5
intervals==0.9.2
itsdangerous==2.0.1
Jinja2==3.0.2
kombu==5.2.1
Mako==1.1.5
MarkupSafe==2.0.1
oauthlib==3.1.1
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.26
SQLAlchemy-Utils==0.37.9
toml==0.10.2
urllib3==1.26.7
URLObject==2.4.3
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
zipp==3.6.0
